requirements.txt
WTForms>=2.2.1,==2.*
#psycopg2-binary>=2.8.0,==2.*
psycopg2>=2.8.0,==2.*
SQLAlchemy>=1.3.0,==1.*
Flask-SQLAlchemy>=2.4.0,==2.*
sqlalchemy-json>=0.4.0,==0.*
Flask-Migrate>=2.5.0,==2.*
alembic>=1.5.0,==1.*
python-dotenv>=0.16.0,==0.*
password_strength>=0.0.3.post2,==0.*
Markdown==3.2.1
email-validator>=1.1.2,==1.*
passlib>=1.7.2,==1.*
Unidecode==1.1.1
Flask>=1.1.1,==1.*
Flask_WTF>=0.14.3,==0.*
Flask_Babel==1.0.0
unicodecsv>=0.14.1,==0.*
beautifulsoup4>=4.9.3,==4.*
flask_session>=0.3.2,==0.*
mjml>=0.6.1,==0.*
Pillow>=8.2.0,==8.*
cryptography>=3.4.7,==3.*
minio>=7.0.3,==7.*
urllib3>=1.26.4,==1.*
gunicorn>=20.1.0,==20.*
flask-marshmallow>=0.14.0,==0.*
marshmallow-sqlalchemy>=0.26.1,==0.*
